Staging refactoring

+++ Assets/Scripts/Models/TimelineWaterData.cs
@@ -15,6 +15,6 @@ namespace Assets.Scripts.Models
     public class PrecipByYear
     {
         public int year { get; set; }
-        public int precipitation { get; set; }
+        public float precipitation { get; set; }
     }
 }
